requirements.txt
python=3.6.5
numpy=1.19.4
pytorch=1.9.1
torchvision=0.10.1
pyro-ppl=1.5.0
matplotlib=3.2.2
scikit-learn=0.23.2
argparse=1.1
pandas=1.1.3
scipy=1.2.0
